Move signature related code into separate module
